Added functionality to the decision tree project
In the decision tree projec it is a TDD project that goes through the flow of creating a user story and then testing it and then implementing it. Each step flows on from the next so for example you can not write your code until you have written your unit tests etc.